Line destroying system adjustement [PART 1]
Some bug fix has been done, moreover several methods refactoring is now done, the project has been migrated to the unity lastest version.
From now on when one line or more is achieved, the system detects it and destroy thoses lines.
However, there is a need to review the pieces colliders settings to fit the new requiements for the line destroying feature.