Line destroying system adjustement [PART 2]
The collider system has been reviewed for this feature, the field size calculation is now working.
The line destruction bug about partial pieces is fixed and a colision bug with the L pattern piece too.
Also some coments has been added in the code to explain mechanics and refactoring for readability purpose.