Game over detection
From now, on the game will be over if the player piece is outside the screen.
Texts will be displayed when the requirement of game over are met.
The player will be able to restart the game from the begining by pressing the enter buton.
Moreover an issue fix has been done (background size calculation)